edit/delete functionality works

+++ db/connectLinkDelete.sql
@@ -1,2 +1,1 @@
-DELETE from user_links where id = ${id};
-SELECT * FROM user_links WHERE uid = (SELECT uid FROM user_links WHERE id= ${id});
+DELETE from user_links where id = ${id} returning uid;
